Context supports multiple principals (subject)
Provision for integration with external entitlement management solutions
like LDAP or JWT. E.g. in the case of JWT the first principal of the 
context would be the subject, and other principals would be subject
roles. This way the NSF-based solution may map low-level permissions to
role principals, and leave mapping of subject to roles to the external
system.